+++ packages/react-core/src/components/Dropdown/Dropdown.tsx
@@ -73,7 +73,7 @@ export const Dropdown: React.FunctionComponent<DropdownProps> = ({
   alignments,
   contextProps,
   menuAppendTo = 'inline',
-  isFlipEnabled = false,
+  isFlipEnabled = true,
   removeFindDomNode = false,
   ...props
 }: DropdownProps) => (