+++ packages/react-core/src/components/Pagination/Pagination.tsx
@@ -177,7 +177,6 @@ const handleInputWidth = (lastPage: number, node: HTMLDivElement) => {
   }
 };
 
-let paginationId = 0;
 export class Pagination extends React.Component<PaginationProps, { ouiaStateId: string }> {
   static displayName = 'Pagination';
   paginationRef = React.createRef<HTMLDivElement>();
@@ -211,7 +210,7 @@ export class Pagination extends React.Component<PaginationProps, { ouiaStateId: 
     itemsStart: null,
     itemsEnd: null,
     perPageOptions: defaultPerPageOptions,
-    widgetId: 'pagination-options-menu',
+    widgetId: 'options-menu',
     onSetPage: () => undefined,
     onPerPageSelect: () => undefined,
     onFirstClick: () => undefined,
@@ -325,7 +324,7 @@ export class Pagination extends React.Component<PaginationProps, { ouiaStateId: 
           isSticky && styles.modifiers.sticky,
           className
         )}
-        id={`${widgetId}-${paginationId++}`}
+        {...(widgetId && { id: `${widgetId}-${variant}-pagination` })}
         {...getOUIAProps(Pagination.displayName, ouiaId !== undefined ? ouiaId : this.state.ouiaStateId, ouiaSafe)}
         {...props}
       >
@@ -366,7 +365,7 @@ export class Pagination extends React.Component<PaginationProps, { ouiaStateId: 
           lastPage={lastPage}
           onPerPageSelect={onPerPageSelect}
           dropDirection={dropDirection}
-          widgetId={widgetId}
+          widgetId={`${widgetId}-${variant}`}
           toggleTemplate={toggleTemplate}
           isDisabled={isDisabled}
           perPageComponent={perPageComponent}