+++ packages/react-core/src/helpers/index.ts
@@ -10,3 +10,4 @@ export * from './useIsomorphicLayout';
 export * from './KeyboardHandler';
 export * from './resizeObserver';
 export * from './useInterval';
+export * from './datetimeUtils';